css/image-fit/js/ua-stylesheet/embed/005.html

<!doctype html>
<title>embed default style, flash</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<embed src=../../support/operatest.swf>
<script>
var t = async_test('load', {timeout:2000});
on_event(window, 'load', function() { t.step(function() {
  var e = document.body.firstChild;
  assert_equals(getComputedStyle(e, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
  t.done();
}); });
</script>

